% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutations_datasets.R
\docType{data}
\name{mutations}
\alias{ACC.mutations}
\alias{ACC.mutations,BLCA.mutations,BRCA.mutations,CESC.mutations,CHOL.mutations,COAD.mutations,GBM.mutations,HNSC.mutations,KICH.mutations,KIPAN.mutations,KIRC.mutations,KIRP.mutations,LAML.mutations,LGG.mutations,LIHC.mutations,LUAD.mutations,LUSC.mutations,OV.mutations,PAAD.mutations,PCPG.mutations,PRAD.mutations,READ.mutations,SARC.mutations,SKCM.mutations,STAD.mutations,STES.mutations,TGCT.mutations,THCA.mutations,UCEC.mutations,UCS.mutations,UVM.mutations}
\alias{BLCA.mutations}
\alias{BRCA.mutations}
\alias{CESC.mutations}
\alias{CHOL.mutations}
\alias{COAD.mutations}
\alias{GBM.mutations}
\alias{HNSC.mutations}
\alias{KICH.mutations}
\alias{KIPAN.mutations}
\alias{KIRC.mutations}
\alias{KIRP.mutations}
\alias{LAML.mutations}
\alias{LGG.mutations}
\alias{LIHC.mutations}
\alias{LUAD.mutations}
\alias{LUSC.mutations}
\alias{OV.mutations}
\alias{PAAD.mutations}
\alias{PCPG.mutations}
\alias{PRAD.mutations}
\alias{READ.mutations}
\alias{SARC.mutations}
\alias{SKCM.mutations}
\alias{STAD.mutations}
\alias{STES.mutations}
\alias{TGCT.mutations}
\alias{THCA.mutations}
\alias{UCEC.mutations}
\alias{UCS.mutations}
\alias{UVM.mutations}
\alias{mutations}
\title{Mutations datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.mutations

BLCA.mutations

BRCA.mutations

CESC.mutations

CHOL.mutations

COAD.mutations

GBM.mutations

HNSC.mutations

KICH.mutations

KIPAN.mutations

KIRC.mutations

KIRP.mutations

LAML.mutations

LGG.mutations

LIHC.mutations

LUAD.mutations

LUSC.mutations

OV.mutations

PAAD.mutations

PCPG.mutations

PRAD.mutations

READ.mutations

SKCM.mutations

STAD.mutations

STES.mutations

SARC.mutations

TGCT.mutations

THCA.mutations

UCEC.mutations

UCS.mutations

UVM.mutations
}
\value{
Data frames with mutations data.
}
\description{
Package provides mutations datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package}. The process is described here: \url{http://mi2-warsaw.github.io/RTCGA.data/}.
Mutations data format is explained here https://wiki.nci.nih.gov/display/TCGA/Mutation+Annotation+Format+(MAF)+Specification. There is extra one column with patients' barcodes.
}
\keyword{datasets}

